### 2020-07-10 #2453 @ramadatta
> Hi @ramadatta, thanks for your submission. It looks like there is a problem with your `paper.md` file. As a start, can you try removing these fields from the header?
>
> ```
> citation_author: Prakki et. al.
> year: 2020
> output: rticles::joss_article
> csl: apa.csl
> journal: JOSS
> ```
>
> (just leave the bibliography field in that section of lines, and `date` is ok too)
>
> As you edit this, you can use this page to test that the paper builds properly: https://whedon.theoj.org

Dear @kyleniemeyer, thank you very much! I have edited `paper.md` according to your advice. The modified `paper.md` builds properly with the Whedon paper preview service.
