### 2020-07-10 55fd741
Edited accordingly as per JOSS Editor
