### 2020-07-10 60cec44
Edited accordingly as per JOSS Editor
