### 2020-07-11 #2473 @whedon
[REVIEW]: CPgeneProfiler: A lightweight R package to profile the Carbapenamase genes from genome assemblies
**Submitting author:** @ramadatta (<a href="http://orcid.org/0000-0002-9254-2557">Prakki Sai Rama Sridatta</a>)
**Repository:** <a href="https://github.com/ramadatta/CPgeneProfiler" target ="_blank">https://github.com/ramadatta/CPgeneProfiler</a>
**Version:** v2.1.1
