### 2020-07-11 #2473 @ramadatta
> @ramadatta - please can you add the flagged DOIs to your paper references

@will-rowe, @boasvdp, @fmaguire Thank you so much for kindly agreeing to review the submission!

@will-rowe - I have added the missing DOIs in the paper now and tested the build.
