### 2020-07-11 3e2090a
Fixed DOI and minor additions in README.md
