+++ paper.bib
@@ -6,7 +6,8 @@
   number={18},
   pages={2938--2940},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btx364}
 }
 
 @Book{ggplot,
@@ -26,7 +27,8 @@
   number={1},
   pages={421},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1186/1471-2105-10-421}
 }
 
 @article{inouye2014srst2,
@@ -37,7 +39,8 @@
   number={11},
   pages={90},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s13073-014-0090-6}
 }
 
 @article{zankari2012identification,
@@ -48,7 +51,8 @@
   number={11},
   pages={2640--2644},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/jac/dks261}
 }
 
 @article{lakin2017megares,
@@ -59,7 +63,8 @@
   number={D1},
   pages={D574--D580},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkw1009}
 }
 
 @article{jia2016card,
@@ -68,7 +73,8 @@
   journal={Nucleic acids research},
   pages={gkw1004},
   year={2016},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkw1004}
 }
 
 @article{gupta2014arg,
@@ -79,7 +85,8 @@
   number={1},
   pages={212--220},
   year={2014},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol},
+  doi={10.1128/aac.01310-13}
 }
 
 @article{hunt2017ariba,
@@ -89,7 +96,8 @@
   volume={3},
   number={10},
   year={2017},
-  publisher={Microbiology Society}
+  publisher={Microbiology Society},
+  doi={10.1099/mgen.0.000131}
 }
 
 @article{hendriksen2019using,
@@ -99,7 +107,8 @@
   volume={7},
   pages={242},
   year={2019},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi={10.3389/fpubh.2019.00242}
 }
 
 @inproceedings{cai2017prevalence,
@@ -109,9 +118,9 @@
   volume={4},
   number={3},
   year={2017},
-  organization={Oxford University Press}
+  organization={Oxford University Press},
+  doi={10.1093/ofid/ofx176}
 }
-
 
 @article{van2013carbapenem,
   title={Carbapenem-resistant Enterobacteriaceae: a review of treatment and outcomes},
@@ -121,7 +130,8 @@
   number={2},
   pages={115--120},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.diagmicrobio.2012.11.009}
 }
 
 @article{papp2011carbapenems,
@@ -132,7 +142,8 @@
   number={11},
   pages={4943--4960},
   year={2011},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol},
+  doi={10.1128/AAC.00296-11}
 }
 
 @article{spencer2019whole,
@@ -143,7 +154,8 @@
   number={3},
   pages={187--199},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jinf.2018.11.003}
 }
 
 @article{tamariz2018draft,
@@ -154,8 +166,10 @@
   number={13},
[…]
