### 2020-07-11 826bce8
minor additions in README.md
