+++ paper.bib
@@ -179,7 +179,7 @@
   number={6},
   year={2019},
   publisher={American Society for Microbiology (ASM)},
-  doi = {10.1128/aac.02282-18}
+  doi={10.1128/aac.02282-18}
 }
 
 
@@ -192,7 +192,7 @@
   pages={S521--S528},
   year={2019},
   publisher={Oxford University Press US},
-  doi = {10.1093/cid/ciz824}
+  doi={10.1093/cid/ciz824}
 }
 
 
