### 2020-07-11 56eb6e5
minor additions in README.md
