+++ paper.bib
@@ -7,7 +7,7 @@
   pages={2938--2940},
   year={2017},
   publisher={Oxford University Press},
-  doi = {10.1093/bioinformatics/btx364}
+  doi={10.1093/bioinformatics/btx364}
 }
 
 @Book{ggplot,
@@ -28,7 +28,7 @@
   pages={421},
   year={2009},
   publisher={Springer},
-  doi = {10.1186/1471-2105-10-421}
+  doi={10.1186/1471-2105-10-421}
 }
 
 @article{inouye2014srst2,
