### 2020-07-11 29eda35
minor additions in README.md
