+++ paper.bib
@@ -17,6 +17,7 @@
     year = {2016},
     isbn = {978-3-319-24277-4},
     url = {https://ggplot2.tidyverse.org},
+    doi={10.1007/978-0-387-98141-3}
   }
 
 @article{camacho2009blast,
