### 2020-07-13 ramadatta/CPgeneProfiler#5 @fmaguire
DESCRIPTION file metadata incomplete
Please add the authorship/ORCID information etc to the R `DESCRIPTION` metadata file.

Currently looks like it is just placeholder metadata.
