### 2020-07-13 b3b3efa
Fix typos in README
