### 2020-07-13 ramadatta/CPgeneProfiler#6 @fmaguire
MANUSCRIPT: Conflation of AMR databases and AMR gene detection tools
```To this end, several freely available bioinformatics resources were identified to find the antimicrobial resistance genes in nucleotide as well as amino acid sequence data [@hendriksen2019using]. Some of these resources include, but not limited to are ARIBA [@hunt2017ariba], ARG-ANNOT [@gupta2014arg], CARD database [@jia2016card], MEGARes [@lakin2017megares], NCBI-AMRFinder (https://www.ncbi.nlm.nih.gov/pathogens/antimicrobial-resistance/AMRFinder/), KmerResistance (https://cge.cbs.dtu.dk/services/KmerResistance/), SRST2 [@inouye2014srst2] and ResFinder [@zankari2012identification].```

ARIBA, NCBI-AMRFinder, KmerResistance, SRST2, and Resfinder, are all the names of tools that can be used to detect AMR genes whereas: CARD, ARG-ANNOT, MEGARes, and confusingly ResFinder (the tool and the database have the same name) are the names of databases of AMR genes.  As several of these tools (e.g. ARIBA and SRST2) can be easily configured to use different databases it is important to distinguish tools and databases.  Additionally, the databases themselves tend to have associated tools they have generated e.g.:

- NCBI-AMRFinder is the a tool developed for the National Database of Antibiotic Resistant Organisms (AMRFinderPlus is the main version now which supports mutational AMR).
- AMRPlusPlus is the tool developed by the MEGARes team for using MEGARes data.
- RGI is the tool developed by the CARD team for detecting AMR.
- KmerResistance/PointFinder/ResFinder are all tools developed by the ResFinder team for detecting genes in their databases.

**Minor Suggested Change**: Please rephrase this paragraph to make the distinction clear between databases and AMR detection tools.
