### 2020-07-13 ramadatta/CPgeneProfiler#7 @fmaguire
INSTALLATION: Add R dependencies to R package so they are installed automatically
Please add the R package dependencies to the DESCRIPTION e.g. dplyr, tidyverse, UpSetR etc so they are automatically installed when the user runs `devtools::install_github("ramadatta/CPgeneProfiler")`

Optional: consider making a bioconda package (conda isn't exclusively python and there are a ton of non-python bioinformatics tools within bioconda now) that will install the tool AND the external BLAST+ dependency: the bioconda [docs](https://bioconda.github.io/contributor/index.html) and a [simple R package that already exists](https://github.com/bioconda/bioconda-recipes/blob/master/recipes/r-rBLAST/meta.yaml) can be used as a template.
