### 2020-07-13 6f7d71d
Fix install instructions
