### 2020-07-13 61b9527
Remove prompt from quickstart commands to facilitate copy and paste
