### 2020-07-13 ab19e2f
Add R package install commands and link to blast+
