### 2020-07-13 ramadatta/CPgeneProfiler#9 @fmaguire
FUNCTIONALITY No assembly name axes labels in heatmap
When run on test data the generated heatmap doesn't contain assembly name axes labels. See example here:
![CPgeneProfile](https://user-images.githubusercontent.com/1698629/87339484-111c1d00-c51d-11ea-99ed-7cdd072c6835.png)

Without this the comparison utility of the heatmap is limited.

The colorbar label should be more specific than "value"
