### 2020-07-13 ramadatta/CPgeneProfiler#10 @fmaguire
FUNCTIONALITY upSet plot contains an extra blank page  
On the testdata the generated upSet plot seems to contain an extra blank page.

[upSetR.pdf](https://github.com/ramadatta/CPgeneProfiler/files/4914396/upSetR.pdf)
