### 2020-07-13 ramadatta/CPgeneProfiler#11 @fmaguire
AUTOMATED TESTS: add specific test section of README
While there is testdata and the quickstart example includes it, you should have a specific "Testing" section in your README that runs the testdata and describes the expected outcome.

Ideally this would include automating this testing process (e.g. Continuous Integration).
