### 2020-07-14 ramadatta/CPgeneProfiler#12 @fmaguire
FUNCTIONALITY Splitting single function into individual functions would greatly improve utility
If there were more granularity than just one master function taking in the fasta+db and generating all the plots this package would improve the utility of this package.  By having the entire program/package one large program only users with the exact same use-case as you will benefit from your work.  If the package was split up into several different functions it would make it a lot more flexible.

Specifically, it is worth creating separate documented functions/APIs for each summary/visualisation/analysis with their own parameters.  For example, a function generating the upset plot could have parameters to just generate the plot for specific CP genes/families.  Similarly for other visualisations/summaries even the BLAST+parsing as a separate function would make it easier for users to tweak the BLAST parameters/change the specifics of the homology search.
