### 2020-07-14 ramadatta/CPgeneProfiler#13 @fmaguire
COMMUNITY GUIDELINES: need added
`Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support`
