### 2020-07-14 ramadatta/CPgeneProfiler#14 @fmaguire
FUNCTIONALITY Lack of active curation/updates for ARG-ANNOT mean cannot be used for CP surveillance 
Need a justification for why ARG-ANNOT was used.
As the author's point out genomic surveillance of CP genes is important, therefore using a database that has not been updated in two years is a problem as it won't capture the latest discovered CP diversity.
For comparison, a crude `grep -c` of different beta-lactamase AMR gene names (mostly CP) shows for most that ARG-ANNOT lacks a lot of CP alleles retained in databases are updated multiple times per year.

| db | version | VIM | OXA | NDM | KPC | IMP | IMI |
|--  |--       |--   |--   |--   |--   |--   |--   |
| CARD | 3.0.9 | 39 | 439 | 27 | 19 | 47 | 6 |
| ResFinder | 2020-06-02 | 53 | 526 | 24 | 39 | 68 | 4 |
| NCBI Bacterial AMR Reference Gene Database | 2020-06-11.1 | 68 | 874 | 29 | 51 | 80 | 18
| **ARG-ANNOT/CPGeneProfiler** | v4 May 2018 | 42 | 33 | 14 | 20 | 39 | 6 |

Highly, highly recommend modifying this tool to use NCBI, CARD, or ResFinder so that its CP repertoire is actively curated and regularly updated.
