### 2020-07-14 ramadatta/CPgeneProfiler#15 @fmaguire
REFERENCES: need added/updated
Generally, these databases and tools are reliant on constant citations of the latest release papers (to prove ongoing engagement/use of the database) for continued funding.

**Minor Suggested Change**: Please cite the latest manuscript for all the tools and databases mentioned e.g.:

- [MEGARes 2.0](https://doi.org/10.1093/nar/gkz1010)
- [KmerResistance](https://doi.org/10.1093/jac/dkw184)
- [NDARO](https://doi.org/10.1093/nar/gkz899)
- [AMRFinder](https://doi.org/10.1128/AAC.00483-19)
- [CARD 2020](https://doi.org/10.1093/nar/gkz935) _ @will-rowe : minor COI here as I am an author on this paper so happy to leave this review comment if that is an issue _
