### 2020-07-14 ramadatta/CPgeneProfiler#16 @fmaguire
STATE OF FIELD: No comparison to AMR detection tools with summary utilities
`Current antimicrobial resistance gene finding tools can detect the genes from either genome assemblies or sequencing reads but do not exclusively report a genetic profile for the presence of CP genes across all the genome assemblies and neither provide the ability to extract and visualize the co-carriage of multiple CP genes. Therefore, to address this need, we describe here a lightweight R package, CPgeneProfiler that scans bacterial genome sequences to detect the presence of CP genes using R framework`

While there are no existing tools that produce these summaries for only CP genes, there are several packages/tools that allow either generation of comparative tables across genomes and some visualisations for a much more comprehensive AMR gene complement e.g.
[sraX](doi.org/10.3389/fmicb.2020.00052) allows generation of numerous navigable resistome comparison visualisations,
[RGI](https://github.com/arpcard/rgi#generating-heat-maps-of-rgi-main-results) has a heatmap feature that allows users to produce similar comparison visualisations across multiple genomes,
and [abricate](https://github.com/tseemann/abricate#combining-reports-across-samples) has a command to generate a summary table from multiple genomes.

These need mentioned for point of comparison and the benefits of the specialisation of CPGeneProfiler carefully articulated.
