### 2020-07-14 ramadatta/CPgeneProfiler#17 @fmaguire
STATEMENT OF NEED: paper needs to explain WHY all the plots are useful for CP surveillance
Need to explain the utility of contig length distribution analysis, co-carriage sets, and so on.

What specifically are these useful for in surveillance/profiling of CP genes? e.g. plasmid/MGE identification etc.
