### 2020-07-14 ramadatta/CPgeneProfiler#18 @fmaguire
MANUSCRIPT: Need to clarify that not all OXA are carbapenemases
e.g. https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5277539/
