### 2020-07-14 ramadatta/CPgeneProfiler#19 @fmaguire
REFERENCES Unclear exactly why the README references are cited
They are too partial to include all the appropriate references (e.g. all the dependencies) and it unclear why those specific ones are included (especially the Bai et. al. 2015 paper).  Please clarify.
