### 2020-07-14 #2473 @fmaguire
## Overview

In this tool the author's provide an R package that generates some potentially useful diagnostic plots for AMR researchers with limited data analysis/visualisation experience and an interest in carbapenemase genes.
Unfortunately, the utility of this tool is currently hampered by its lack of modularity, choice of AMR database, and limited ease of installation/dependency management.
While there are no existing AMR detection tools that do exactly the same summary/analysis that this package provides, there are several that provide more comprehensive AMR profiling (more than just carbapenemases), some visualisation, and generation of outputs that can be visualised as the analyst sees fit with relatively little work.

With improved documentation, a fairly major overhaul to improve modularity and packaging (listed below) CPgeneProfiler could represent a handy set of visualisation utilities for this targeted research/public health group.
However, as it currently stands CPgeneProfiler isn't widely applicable outside the specific internal analytical use-case for which it was developed.

## Installation

- Minor: Few minor typos in README and installation instructions, fixed in [PR](https://github.com/ramadatta/CPgeneProfiler/pull/20)
- Major: Installation procedure needs streamlining (using dependency management to handle installation of required R packages): [issue](https://github.com/ramadatta/CPgeneProfiler/issues/7)

## Functionality

- Major: Generated plots don't contain assembly names so main utility of comparison of CP genes across assemblies is limited until this is fixed e.g. heatmap y-axis labels: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/9)
- Major: Entire package is a single function greater limiting utility/flexibility: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/12)
- Major: Built around ARG-ANNOT (not hard to change), no updates/curation in >2 years therefore CPGeneProfiler cannot be used for surveillance of many novel CP alleles: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/14)
- Minor: Artefacts and unspecific labels in some plots: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/10) also mentioned in [issue](https://github.com/ramadatta/CPgeneProfiler/issues/9)
- Minor: Typo in output filename: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/8)

## Documentation

- Major: R package DESCRIPTION file metadata needs completed: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/5)
- Minor: README citations need fixed: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/19)
- Minor: Example usage is sufficient although could be further fleshed out.
- Minor: Test data but no automated tests: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/11)
- Minor: No community guidelines: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/13)

## Software Paper

- Major: Paper and documentation doesn't really explain the specific utility of each output of tool: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/17)
- Minor: Should clarify that not all OXA beta-lactamases are CP genes: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/18)
- Minor: Manuscript conflates AMR databases and AMR gene detection tools: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/6)
- Minor: References need added/updated for AMR databases/tools: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/15)
- Minor: Comparison to more comprehensive AMR detection tools with summary plots etc needs added and benefits of CPgeneProfiler's specialisation clearly stated: [issue](https://github.com/ramadatta/CPgeneProfiler/issues/16)
