### 2020-07-14 ramadatta/CPgeneProfiler#15 @will-rowe
I think this is a valid point and I see no issue with CARD being cited here.
