### 2020-07-14 #2473 @will-rowe
Thanks @fmaguire - that's an excellent review.

@ramadatta - feel free to begin working on these points now. @boasvdp will also be providing feedback so please keep an eye out for that too.

I'm due to go on annual leave for a week next week. I'll check back in here when I'm back but please ping another editor if you need assistance next week.
