### 2020-07-26 ramadatta/CPgeneProfiler#21 @boasvdp
FUNCTIONALITY: Inclusion of a summary report
Including a summary report in the output would be very helpful, especially for researchers without bioinformatics expertise (please ignore if this is not the aim of the tool).

Ideally, the summary report would combine information that is scattered across different output files now. For example this could include contig name, contig position of the CP gene, contig length, co-occurring CP genes, blast results (coverage, nucl identity, e-value), etc.
