### 2020-07-26 ramadatta/CPgeneProfiler#22 @boasvdp
FUNCTIONALITY: Adding option for output location
Currently, output is automatically moved to a folder in the input directory. It would be helpful if the output directory could be specified in the `CPgeneProfiler()` function.
