### 2020-07-26 ramadatta/CPgeneProfiler#23 @boasvdp
FUNCTIONALITY: Blastn thresholds for CP gene identification
The threshold for gene identification (if I understood correctly) is now an e-value score of 1e-6, and then selecting the best hit at a location. Other tools usually employ thresholds based on % nucleotide identity and % coverage, which is also a lot clearer for non-bioinformaticians.

Adding an option which can filter based on nucleotide identity or coverage (or at least reporting these clearly) would improve the tool
