### 2020-07-26 ramadatta/CPgeneProfiler#23 @boasvdp
Upon closer inspection, I see that the filtering step excludes all hits that have less than 100% nucleotide identity and coverage. Although most common CP genes will be identified using this approach, one would expect false positives if genes are only partially assembled or new alleles are found. Being able to manage the thresholds would improve sensitivity
