### 2020-07-26 ramadatta/CPgeneProfiler#24 @boasvdp
DOCUMENTATION: No source provided for the delineation of CP genes
Especially for the blaOXA gene family, how do you decide which genes are carbapenemase genes and which are not? Is there  a citeable source which you can provide for this?

Ideally this would be a source which is continuously updated, which allows identification of recently emerged alleles
