### 2020-07-26 ramadatta/CPgeneProfiler#25 @boasvdp
DOCUMENTATION: Default blastn settings are not indicated
CPgeneProfiler currently only looks for exact matches of CP genes (100% nucleotide identity, 100% coverage), which is not clear from the documentation.
