### 2020-07-26 #2473 @boasvdp
# General
The described tool identifies carbapenemase genes from assembled bacterial genomes and produces additional plots, especially aimed at co-carriage. The identification of carbapenemase genes is possible through other tools, but I am unaware of any tools which produce the additional plots that are included in the CPgeneProfiler output. The tool is installable (although this could be improved) and does what it promises. However, improvements are possible and in some cases necessary.

One major outstanding question for me is the 'audience' for this software. What level of bioinformatics expertise would you expect for people using your tool? Is this a tool for bioinformaticians, or is this intended for molecular biologists with some R experience? This influences some of the suggestions I have made below, as some will apply more or less depending on the bioinformatics expertise of the audience you're aiming for.

I support all notions made by @fmaguire above, and will try to provide additional feedback where needed.

# Installation
- I second the @fmaguire's idea to make a (bio)conda package, as all dependencies are already available in conda. Otherwise, automated handling of dependencies would be good.

# Functionality
- Summary report, especially helpful for researchers without a lot of bioinformatics expertise: https://github.com/ramadatta/CPgeneProfiler/issues/21
- As @fmaguire noted, having the possibility to run the tool in a modular way would be helpful (if the tool is aimed at bioinformaticians)
- The addition of some extra options would improve the tool, esp. regarding output (https://github.com/ramadatta/CPgeneProfiler/issues/22) and blastn thresholds (https://github.com/ramadatta/CPgeneProfiler/issues/23)
- ARG-ANNOT is not regularly updated, using a better database (or ideally multiple databases) would improve the tool. NCBI has a regularly updated database: https://www.ncbi.nlm.nih.gov/pathogens/isolates#/refgene/

# Documentation
- There is no source provided for the delineation of carbapenemase genes: https://github.com/ramadatta/CPgeneProfiler/issues/24
- There is no clear mention of the default settings: https://github.com/ramadatta/CPgeneProfiler/issues/25

# Software paper
- Currently, there is no comparison with other tools. Due to the strict treshold used by CPgeneProfiler, this could lead to a lower sensitivity, esp. with misassemblies/sequencing errors/new alleles. If custom thresholds are added as an option, I would not expect too many differences with other tools as these are mostly based on blast as well. I have not opened an issue for this one as I expect this will resolve itself.
