### 2020-07-27 #2473 @ramadatta
@boasvdp - Thank you so much for the comments!

I actually intended to make this package helpful to biologists who have only little experience in R. I am working on the package currently and trying to make it modular. I also appreciate both of your suggestions on changing the database and working on this. I am making an effort to make the functionality more flexible. Thanks again!
