+++ paper.bib
@@ -1,0 +1,202 @@
+@article{conway2017upsetr,
+  title={UpSetR: an R package for the visualization of intersecting sets and their properties},
+  author={Conway, Jake R and Lex, Alexander and Gehlenborg, Nils},
+  journal={Bioinformatics},
+  volume={33},
+  number={18},
+  pages={2938--2940},
+  year={2017},
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btx364}
+}
+
+@Book{ggplot,
+    author = {Hadley Wickham},
+    title = {ggplot2: Elegant Graphics for Data Analysis},
+    publisher = {Springer-Verlag New York},
+    year = {2016},
+    isbn = {978-3-319-24277-4},
+    url = {https://ggplot2.tidyverse.org},
+    doi={10.1007/978-0-387-98141-3}
+  }
+
+@article{camacho2009blast,
+  title={BLAST+: architecture and applications},
+  author={Camacho, Christiam and Coulouris, George and Avagyan, Vahram and Ma, Ning and Papadopoulos, Jason and Bealer, Kevin and Madden, Thomas L},
+  journal={BMC bioinformatics},
+  volume={10},
+  number={1},
+  pages={421},
+  year={2009},
+  publisher={Springer},
+  doi={10.1186/1471-2105-10-421}
+}
+
+@article{inouye2014srst2,
+  title={SRST2: rapid genomic surveillance for public health and hospital microbiology labs},
+  author={Inouye, Michael and Dashnow, Harriet and Raven, Lesley-Ann and Schultz, Mark B and Pope, Bernard J and Tomita, Takehiro and Zobel, Justin and Holt, Kathryn E},
+  journal={Genome medicine},
+  volume={6},
+  number={11},
+  pages={90},
+  year={2014},
+  publisher={Springer},
+  doi={10.1186/s13073-014-0090-6}
+}
+
+@article{zankari2012identification,
+  title={Identification of acquired antimicrobial resistance genes},
+  author={Zankari, Ea and Hasman, Henrik and Cosentino, Salvatore and Vestergaard, Martin and Rasmussen, Simon and Lund, Ole and Aarestrup, Frank M and Larsen, Mette Voldby},
+  journal={Journal of antimicrobial chemotherapy},
+  volume={67},
+  number={11},
+  pages={2640--2644},
+  year={2012},
+  publisher={Oxford University Press},
+  doi={10.1093/jac/dks261}
+}
+
+@article{lakin2017megares,
+  title={MEGARes: an antimicrobial resistance database for high throughput sequencing},
+  author={Lakin, Steven M and Dean, Chris and Noyes, Noelle R and Dettenwanger, Adam and Ross, Anne Spencer and Doster, Enrique and Rovira, Pablo and Abdo, Zaid and Jones, Kenneth L and Ruiz, Jaime and others},
+  journal={Nucleic acids research},
+  volume={45},
+  number={D1},
+  pages={D574--D580},
+  year={2017},
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkw1009}
+}
+
+@article{jia2016card,
+  title={CARD 2017: expansion and model-centric curation of the comprehensive antibiotic resistance database},
+  author={Jia, Baofeng and Raphenya, Amogelang R and Alcock, Brian and Waglechner, Nicholas and Guo, Peiyao and Tsang, Kara K and Lago, Briony A and Dave, Biren M and Pereira, Sheldon and Sharma, Arjun N and others},
+  journal={Nucleic acids research},
+  pages={gkw1004},
+  year={2016},
[…]

[…]
