+++ R/CPgeneProfiler.R
@@ -1,5 +1,23 @@
 #' Generate a profile of carbapenamase genes from the genome assemblies
 #'
+<<<<<<< HEAD
+#' `CPgeneProfiler` checks for a list of CarbaPenamase (CP) genes from a list of
+#' input genome assemblies provided in fasta file format. It reports the profile of all
+#' the CP genes available in the genome assemblies in the format of simple heatmap.
+#' Apart from this, it also reports the presence of cocarriage of CP genes across the
+#' input genome assemblies. Other assembly statistics such as N50, N90, Assembly Size from the
+#' assembly are calculated and plots of length distribution of CP gene contigs from
+#' the list of assemblies are reported. The following functions are used to run
+#' CPgeneProfiler: `cpblast()`,`filt_blast`,`cocarriage()`,`cpprofile()`,`upsetR_plot()`,
+#' `plot_conlen()`,`assemblystat()`, `db_summary()`, `cp_summarize()`
+#'
+#' @docType package
+#' @name CPgeneProfiler
+NULL
+
+
+
+=======
 #' CPgene-profiler checks for a list of CarbaPenamase (CP) genes from a list of
 #' genome assemblies provided in fasta file format. It reports the profile of all
 #' the CP genes available in the genome assemblies in the format of simple heatmap.
@@ -738,3 +756,4 @@ CPgeneProfiler <- function(fastafiles_location,db_location){
   print(new) # print in nice format
   # ELAPSED TIME
 }
+>>>>>>> 8b4803c980595a95edd1ca1e46a9633c509f109b

[…]
