+++ DESCRIPTION
@@ -1,6 +1,5 @@
 Package: CPgeneProfiler
-<<<<<<< HEAD
-Title: Profiles the Carbapenamase genes from the genome assemblies 
+Title: Profiles the Carbapenamase genes from the genome assemblies (One Line, Title Case)
 Version: 2.1.0.9000
 Authors@R: 
     c(person(given = "Sai Rama Sridatta",
@@ -43,25 +42,9 @@ Imports: tidyverse (>= 1.3.0),
 	 pdftools,
 	grid
 License: GPL-2
-=======
-Title: What the Package Does (One Line, Title Case)
-Version: 0.0.0.9000
-Authors@R: 
-    person(given = "First",
-           family = "Last",
-           role = c("aut", "cre"),
-           email = "first.last@example.com",
-           comment = c(ORCID = "YOUR-ORCID-ID"))
-Description: What the package does (one paragraph).
-License: `use_mit_license()`, `use_gpl3_license()` or friends to
-    pick a license
->>>>>>> 8b4803c980595a95edd1ca1e46a9633c509f109b
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.0
-<<<<<<< HEAD
 URL: https://github.com/ramadatta/CPgeneProfiler
 BugReports: https://github.com/ramadatta/CPgeneProfiler/issues
-=======
->>>>>>> 8b4803c980595a95edd1ca1e46a9633c509f109b
