+++ R/CPgeneProfiler.R
@@ -1,7 +1,5 @@
 #' Generate a profile of carbapenamase genes from the genome assemblies
 #'
-<<<<<<< HEAD
-<<<<<<< HEAD
 #' `CPgeneProfiler` checks for a list of CarbaPenamase (CP) genes from a list of
 #' input genome assemblies provided in fasta file format. It reports the profile of all
 #' the CP genes available in the genome assemblies in the format of simple heatmap.
@@ -18,748 +16,3 @@ NULL
 
 
 
-=======
-=======
->>>>>>> 8b4803c980595a95edd1ca1e46a9633c509f109b
-#' CPgene-profiler checks for a list of CarbaPenamase (CP) genes from a list of
-#' genome assemblies provided in fasta file format. It reports the profile of all
-#' the CP genes available in the genome assemblies in the format of simple heatmap.
-#' Apart from this, it also reports the presence of co-carriage of CP genes within
-#' an assembly. Other assembly statistics such as N50, N90, Assembly Size from the
-#' assembly are calculated and plots of length distribution of CP gene contigs from
-#' the list of assemblies are reported.
-#'
-#' @param fastafiles_location Need folder location with all the fasta files to check for CP genes
-#' @param db_location Need folder location of the downloaded ARG-annot_CPGene_DB.fasta file
-#' @export
-
-CPgeneProfiler <- function(fastafiles_location,db_location){
-  
-  fastalocation <- fastafiles_location
-  dblocation <- db_location
-  
-  setwd(dblocation)
-  getwd()
-  
-  # Create a BLAST database
-  makeblastdb = "makeblastdb"
-  system2(
-    command = makeblastdb,
-    args = c(
-      "-in",
-      'ARG-annot_CPGene_DB.fasta',
-      "-dbtype",
-      "nucl",
-      "-parse_seqids",
-      "-out",
-      'ARG-annot_CPGene.DB',
-      "-logfile",
-      "log"
-    ),
-    wait = TRUE,
-  )
-  blast_db_location = getwd() # for later use
-  
-  ############--------GO TO ASSEMBLY FILES LOCATION------##########
-  
-  setwd(fastalocation)
-  getwd()
-  
-  ############--------CLEANING IF PREVIOUS OUTPUT FOLDER EXISTS------##########
-  
-  if (dir.exists("CPgeneProfiler_Output")) {
-    unlink("CPgeneProfiler_Output", recursive = TRUE)
-  } else {
-    #print("Dir already exists!")
-  }
-  
-  ############--------VARIABLES------##########
-  
-  blast_db = paste(blast_db_location, "ARG-annot_CPGene.DB", sep = "/")
-  blastn = "blastn"
-  evalue = 1e-6
-  format = "\'6 qseqid sseqid pident nident length mismatch gapopen qstart qend sstart send evalue bitscore qlen slen\'"
-  num_threads <-
-    system2(
-      command = "nproc",
-      args =  c("--all"),
-      wait = TRUE,
-      stdout = TRUE
-    )
-  
-  colnames <- c(
-    "assemblyName",
-    "qseqid",
-    "sseqid",
-    "pident",
-    "nident",
-    "length",
-    "mismatch",
-    "gapopen",
-    "qstart",
-    "qend",
-    "sstart",
-    "send",
-    "evalue",
-    "bitscore",
-    "qlen",
-    "slen"
-  )
-  
-  
-  
-  ############--------CALCULATING TIME------##########
-  old <- Sys.time()
-  
-  ############--------1) BLASTN ASSEMBLY ON ARG-ANNOT DATA------##########
-  
-  #List of all the fasta files
[…]
