+++ man/png_merge.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/png_merge.R
+\name{png_merge}
+\alias{png_merge}
+\title{Merge png images to pdf for summary report}
+\usage{
+png_merge(pdfFile, pngFiles)
+}
+\description{
+Function used for summarizing all the results from CPgeneProfiler output
+}
