+++ man/CPgeneProfiler.Rd
@@ -1,22 +1,17 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/CPgeneProfiler.R
+\docType{package}
 \name{CPgeneProfiler}
 \alias{CPgeneProfiler}
 \title{Generate a profile of carbapenamase genes from the genome assemblies}
-\usage{
-CPgeneProfiler(fastafiles_location, db_location)
-}
-\arguments{
-\item{fastafiles_location}{Need folder location with all the fasta files to check for CP genes}
-
-\item{db_location}{Need folder location of the downloaded ARG-annot_CPGene_DB.fasta file}
-}
 \description{
-CPgene-profiler checks for a list of CarbaPenamase (CP) genes from a list of
-genome assemblies provided in fasta file format. It reports the profile of all
+\code{CPgeneProfiler} checks for a list of CarbaPenamase (CP) genes from a list of
+input genome assemblies provided in fasta file format. It reports the profile of all
 the CP genes available in the genome assemblies in the format of simple heatmap.
-Apart from this, it also reports the presence of co-carriage of CP genes within
-an assembly. Other assembly statistics such as N50, N90, Assembly Size from the
+Apart from this, it also reports the presence of cocarriage of CP genes across the
+input genome assemblies. Other assembly statistics such as N50, N90, Assembly Size from the
 assembly are calculated and plots of length distribution of CP gene contigs from
-the list of assemblies are reported.
+the list of assemblies are reported. The following functions are used to run
+CPgeneProfiler: \code{cpblast()},\code{filt_blast},\code{cocarriage()},\code{cpprofile()},\code{upsetR_plot()},
+\code{plot_conlen()},\code{assemblystat()}, \code{db_summary()}, \code{cp_summarize()}
 }
