+++ man/cp_summarize.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/cp_summarize.R
+\name{cp_summarize}
+\alias{cp_summarize}
+\title{CPgeneProfiler Results Summary}
+\usage{
+cp_summarize(
+  outdir = "CPgeneProfiler_Output",
+  report = "SummaryPlots",
+  image = "png"
+)
+}
+\arguments{
+\item{outdir}{output directory (default: CPgeneProfiler_Output)}
+
+\item{report}{PDF Summary of all the plots (default: Summary)}
+
+\item{image}{format of all the output image plots. (default:png) Note: All the image plots should be in same format (png/tiff/jpeg) to form a summary report plot file in PDF}
+}
+\description{
+Summarize all the results from CPgeneProfiler output
+}
+\examples{
+
+cp_summarize()
+cp_summarize(outdir = "CPgeneProfiler_Output", report="Summary" , image = "png")
+}
