+++ man/png2pdf.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/png2pdf.R
+\name{png2pdf}
+\alias{png2pdf}
+\title{Converts png images to pdf for summary report}
+\usage{
+png2pdf(outdir, report)
+}
+\description{
+Summarize all the results from CPgeneProfiler output
+}
