+++ man/pdfs2pdf.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/pdfs2pdf.R
+\name{pdfs2pdf}
+\alias{pdfs2pdf}
+\title{Converts multiple pdf files with plots to single pdf for summary report}
+\usage{
+pdfs2pdf(outdir = "CPgeneProfiler_Output", report = "Summary")
+}
+\description{
+Summarize all the plots from CPgeneProfiler output
+}
