+++ NAMESPACE
@@ -1,3 +1,21 @@
 # Generated by roxygen2: do not edit by hand
 
-export(CPgeneProfiler)
+export(CPcontlen)
+export(assemblystat)
+export(cocarriage)
+export(cp_summarize)
+export(cpblast)
+export(cpcore)
+export(cpprofile)
+export(db_summary)
+export(files2move)
+export(filt_blast)
+export(jpeg2pdf)
+export(jpeg_merge)
+export(pdfs2pdf)
+export(plot_conlen)
+export(png2pdf)
+export(png_merge)
+export(tiff2pdf)
+export(tiff_merge)
+export(upsetR_plot)
