### 2020-08-10 99c6031
Removing duplicate copy
