+++ Example_Output/2_Cocarriage_Results/SameCP_SameContig.txt
@@ -1,11 +1,0 @@
-assemblyName qseqid sseqid qlen slen qstart qend length pident cov
-genome_004.fasta 3_length=66727_depth=0.76x OXA-181 66727 998 49850 50847 998 100 100
-genome_004.fasta 3_length=66727_depth=0.76x OXA-181 66727 998 43441 44438 998 100 100
-genome_004.fasta 3_length=66727_depth=0.76x OXA-181 66727 998 37032 38029 998 100 100
-genome_005.fasta 2_length=79441_depth=2.21x_circular=true KPC-2 79441 918 11390 12307 918 100 100
-genome_005.fasta 2_length=79441_depth=2.21x_circular=true KPC-2 79441 918 3810 4727 918 100 100
-genome_012.fasta 2_length=79441_depth=2.21x_circular=true KPC-2 79441 918 11390 12307 918 100 100
-genome_012.fasta 2_length=79441_depth=2.21x_circular=true KPC-2 79441 918 3810 4727 918 100 100
-genome_015.fasta 3_length=66727_depth=0.76x OXA-181 66727 998 49850 50847 998 100 100
-genome_015.fasta 3_length=66727_depth=0.76x OXA-181 66727 998 43441 44438 998 100 100
-genome_015.fasta 3_length=66727_depth=0.76x OXA-181 66727 998 37032 38029 998 100 100
