+++ README.md
@@ -1,10 +1,15 @@
+---
+output:
+  html_document: default
+  pdf_document: default
+---
 # CPgeneProfiler
 Generate a profile of carbapenamase genes from the genome assemblies
 
 ## Author
 [Prakki Sai Rama Sridatta](https://twitter.com/prakki_rama)
 
-## Synopsis
+##### **Synopsis**
 
 1) **CPgeneProfiler** package checks for a list of CarbaPenamase (CP) genes from a list of
  genome assemblies provided in fasta file format. The CP genes are derived from ARG-annot
@@ -21,35 +26,10 @@ Generate a profile of carbapenamase genes from the genome assemblies
  the list of assemblies are reported.  
  
  Currently the package works only on Unix systems.
-
-  ## Quick Start
-  
-  A detailed step-by-step guide for the installation of dependencies and `CPgeneProfiler` is available in this [wiki page](https://github.com/ramadatta/CPgeneProfiler/wiki/Step-by-Step-Guide). 
-
-### Step 1: Download CP gene database using R
-```
-# Specify CP gene database URL 
-> url <- "https://raw.githubusercontent.com/ramadatta/CPgene-profiler
-/master/ARG-annot_CPGene_DB.fasta"
-
-# Specify destination where CP gene database file should be saved 
-> path <- "/home/user/db" # Can change to prefarable location
-> setwd(path)
-> destfile <- "ARG-annot_CPGene_DB.fasta"
-
-# Download the CP gene database file to the folder set in "path"
-> download.file(url, destfile)
-```
-### Step 2: Run the package
-```
-% CPgeneProfiler("/path/Multiple_FastaFiles_Location/","/home/user/db/")
-```
-
-## Prior to CPgeneProfiler installation
+ 
+##### **Prior to CPgeneProfiler installation**
 
 - The following packages are supposed to be installed in R:
-
-  dplyr,
 	 tidyverse,
 	 UpSetR,
 	 scales,
@@ -64,70 +44,150 @@ Generate a profile of carbapenamase genes from the genome assemblies
     
     Note: BLAST version 2.9.0+ was used for the present program although other BLAST+ similar to version 2.9.0+ parameters might also run without problems
     
+##### **Installation**
 
-## CPgeneProfiler Installation
+The R package is available through github repository can be installed using devtools.
 
-Install devtools first, and then need to load them:
 ```
-> install.packages(devtools)
-> library(devtools)
-```
-
-Then install the CPgeneProfiler package using devtools. While installing, switching off to upgrade dependencies
-```
+install.packages("devtools")
 devtools::install_github("ramadatta/CPgeneProfiler")
 ```
 
-## Source
-```
-https://github.com/ramadatta/CPgeneProfiler
-```
-
-# Check installation
+##### **Check installation**
 Ensure you have installed the package properly:
 ```
 ?CPgeneProfiler
 ```
-# Carbapenamase Gene Profiling
 
-## Input Requirements
-* A folder with multiple FASTA files (can be in multiple contigs)
+##### **Input Requirements**
+* Path of a directory with multiple FASTA files (can be in multiple contigs) 
+* Path of Carbapenamase Gene Database directory
 
-## Output Files
+### **Usage**
 
-* A folder "CPgeneProfiler_Output" with the following files
+##### **Step 1: Download CP gene database**
+
+Go to the UNIX/Linux command line terminal and download the `db` folder with SVN
+
+```
+svn export https://github.com/ramadatta/CPgeneProfiler/trunk/testData/db
+```
+else simply [Click](https://downgit.github.io/#/home?url=https://github.com/ramadatta/CPgeneProfiler/tree/master/testData/db) to save database folder and uncompress the `db.zip` folder.
+
[…]
