+++ README.md
@@ -167,28 +167,28 @@ db_summary()
 
 * A folder "CPgeneProfiler_Output" with the following files in respective directories
 
-File | Description
+Command | File | Description
 ----------|--------------
-assemblyStats.txt | A simple text file with N50, N90, Assembly Size for each assembly
-blastResults.filt.txt | Filtered blast results with contains contigs matching CPgenes (default: 100% identity and 100% coverage)
-blastResults.txt | Blast Results of contigs against the CP genes
-Co-carriage_Report.txt | Information of Number of assemblies with the co-carriage broken down to category
-CPContigSizeDist.txt | Contig Size distribution
-CPgeneProfile.png | Carbapenamase Gene Profile (default:png)
-cp_presence-absence_matrix.csv | Presence-absence matrix of CP genes across assemblies
-N50_N90.pdf | N50, N90 vs Assembly Size plot
-"CPGene"_Contig_Dist.png | CP gene contig length distribution (for each CP gene a seperate disribution plot is generated. Default image format: png)
-upset_plot.pdf | Set intersection plot of CP genes across all the input genome assemblies
-DiffCP_DiffContig.txt | Information of assemblies with different CP genes present in different contigs
-DiffCP_SameContig.txt | Information of assemblies with different CP genes present in same contigs 
-SameCP_DiffContig.txt | Information of assemblies with same CP genes present in different contigs
-SameCP_SameContig.txt | Information of assemblies with same CP genes present in same contigs
-SummaryPlots.pdf | All the plots in a single pdf file
+cpblast() | blastResults.txt | Blast Results of contigs against the CP genes
+filt_blast() | blastResults.filt.txt | Filtered blast results with contains contigs matching CPgenes (default: 100% identity and 100% coverage)
+cocarriage() | Co-carriage_Report.txt | Information of Number of assemblies with the co-carriage broken down to category
+cocarriage() | DiffCP_DiffContig.txt | Information of assemblies with different CP genes present in different contigs
+cocarriage() | DiffCP_SameContig.txt | Information of assemblies with different CP genes present in same contigs 
+cocarriage() | SameCP_DiffContig.txt | Information of assemblies with same CP genes present in different contigs
+cocarriage() | SameCP_SameContig.txt | Information of assemblies with same CP genes present in same contigs
+cpprofile() | CPgeneProfile.png | Carbapenamase Gene Profile (default:png)
+plot_conlen() | CPContigSizeDist.txt | Contig Size distribution
+plot_conlen() | "CPGene"_Contig_Dist.png | CP gene contig length distribution (for each CP gene a seperate disribution plot is generated. Default image format: png)
+upsetR_plot() | cp_presence-absence_matrix.csv | Presence-absence matrix of CP genes across assemblies
+upsetR_plot() | upset_plot.pdf | Set intersection plot of CP genes across all the input genome assemblies
+assemblystat() | N50_N90.pdf | N50, N90 vs Assembly Size plot
+assemblystat() | assemblyStats.txt | A simple text file with N50, N90, Assembly Size for each assembly
+cp_summarize() | SummaryPlots.pdf | All the plots in a single pdf file
 
 ## A few example plots created with `CPgeneProfiler`
[…]
