+++ README.md
@@ -90,8 +90,60 @@ Ensure you have installed the package properly:
 ```
 
 ### **Example Usage**
+ 
+`CPgeneProfiler` package can be run using the following functions: `cpblast()`, `filt_blast()`, `cocarriage()`, `cpprofile()`, `upsetR_plot()`, `plot_conlen()`,`assemblystat()`, `cp_summarize()`, `db_summary()`. Below are the examples of `CPgeneProfiler` commands usage. 
 
-##### **Step 1: Download CP gene database**
+
+- Generates NCBI BLAST Results by aligning input genome assemblies against Carbapenamase (CP) gene database
+
+```
+cpblast(fastalocation = "/home/user/CPgeneProfiler/testData/fasta", dblocation = "/home/user/CPgeneProfiler/testData/db")
+
+```
+- Filtering NCBI BLAST Results based on CP gene coverage and percent identity
+
+```
+filt_blast(cpgcov = 100, cpgpident = 100)
+```
+
+- Report cocarriage genes across all the input genome assemblies
+
+```
+cocarriage(cpgcov = 100, cpgpident = 100)
+```
+
+- Generate CP gene Profile
+
+```
+cpprofile()
+```
+
+- Plot CP gene contig length distributions across all the input genome assemblies
+```
+plot_conlen()
+```
+
+- Generate Basic assembly statistics
+
+```
+assemblystat("/home/user/CPgeneProfiler/testData/fasta")
+```
+
+- Summarize the plots and organize all the output files
+
+```
+cp_summarize()
+```
+
+- Database information
+
+```
+db_summary()
+```
+
+### **Working with `testData`**
+
+#### **CP gene Database Download (REQUIRED)**
 
 Go to the UNIX/Linux command line terminal and download the `db` folder with SVN
 
@@ -100,7 +152,7 @@ svn export https://github.com/ramadatta/CPgeneProfiler/trunk/testData/db
 ```
 else simply [Click](https://downgit.github.io/#/home?url=https://github.com/ramadatta/CPgeneProfiler/tree/master/testData/db) to save database folder and uncompress the `db.zip` folder.
 
-##### **Step 2: Download test input data**
+#### **Test Data Download (OPTIONAL)**
 
 To test the package with test input data, Go to the UNIX/Linux command line terminal and download the `fasta` folder with SVN
 
@@ -109,13 +161,8 @@ svn export https://github.com/ramadatta/CPgeneProfiler/trunk/testData/fasta
 ```
 else simply [Click](https://downgit.github.io/#/home?url=https://github.com/ramadatta/CPgeneProfiler/tree/master/testData/fasta) to save fasta folder and uncompress the `fasta.zip` folder.
 
-##### **Step 3: Run the package**
-
-`CPgeneProfiler` package can be run using the following functions: `cpblast()`, `filt_blast()`, `cocarriage()`, `cpprofile()`, `upsetR_plot()`, `plot_conlen()`,`assemblystat()`, `cp_summarize()`, `db_summary()`
-
 
 ##### **Step 3a: A simple NCBI BLAST using `cpblast()` command**
-
 
[…]
