### 2020-08-19 cc80e01
Modular R package
