+++ README.md
@@ -106,34 +106,34 @@ cpblast(fastalocation = "/home/user/CPgeneProfiler/testData/fasta", dblocation =
 filt_blast(cpgcov = 100, cpgpident = 100)
 ```
 
-- Report cocarriage genes across all the input genome assemblies
+- Report cocarriage of CP genes across all the input genome assemblies
 
 ``` r
 cocarriage(cpgcov = 100, cpgpident = 100)
 ```
 
-- Generate CP gene Profile
+- Generate CP gene Profile across all the input genome assemblies
 
 ``` r
-cpprofile()
+cpprofile(xlab = "Carbapenamase Genes", ylab = "Assembly", title = "Carbapenamase Gene Profile Heatmap")
 ```
 
 - Plot CP gene contig length distributions across all the input genome assemblies
 
 ``` r
-plot_conlen()
+plot_conlen(outputType = "png", xlab = "Contig Length", ylab = "Number of Contigs", title = " Contig Length Distribution", colorfill = "#F99245")
 ```
 
-- Generate Basic assembly statistics
+- Generate basic assembly statistics such N50, N90 and Assembly Size and plots comparing Assembly Size with N50, N90 stats
 
 ``` r
-assemblystat("/home/user/CPgeneProfiler/testData/fasta")
+assemblystat(fastalocation = "/home/user/CPgeneProfiler/testData/fasta")
 ```
 
-- Summarize the plots and organize all the output files
+- Summarize the plots and organize all the output files in specific folders
 
 ``` r
-cp_summarize()
+cp_summarize(outdir_loc = "/home/user/Desktop", outdir = "CPgeneProfiler_Output")
 ```
 
 - Database information
@@ -163,27 +163,25 @@ svn export https://github.com/ramadatta/CPgeneProfiler/trunk/testData/fasta
 else simply [Click](https://downgit.github.io/#/home?url=https://github.com/ramadatta/CPgeneProfiler/tree/master/testData/fasta) to save fasta folder and uncompress the `fasta.zip` folder.
 
 
-##### **Step 3a: A simple NCBI BLAST using `cpblast()` command**
+##### **Step 1: A simple NCBI BLAST using `cpblast()` command**
 
 As a first step, CPgeneProfiler generates NCBI BLAST Results by aligning input genome assemblies against Carbapenamase (CP) gene database. Now that you already have a directory with fasta files (should have extensions `.fasta` or `.fa`) in `fasta` folder and cp gene database sequence in `db` folder, you can specify the path of both directories as an input and run the package with `cpblast()` command.
 
 ``` r
-cpblast(fastalocation = "/home/user/CPgeneProfiler/testData/fasta",dblocation = "/home/user/CPgeneProfiler/testData/db",num_threads = 4,evalue = "1e-3")
+cpblast(fastalocation = "/home/user/CPgeneProfiler/testData/fasta", dblocation = "/home/user/CPgeneProfiler/testData/db", num_threads = 4, evalue = "1e-3")
 ```
-The users can adjust BLAST parameters `num_threads` and `evalue` accordingly. If not adjusted, the package runs with default parameters (`num_threads` = 8, `evalue` = "1e-6")
+The users can adjust BLAST parameters `num_threads`, `evalue`, `word_size` and `max_target_seqs` accordingly. If not adjusted and command is simply executed with the file locations for `fasta` and `db`, then default parameters are used for the analysis.
 
 
-##### **Step 3b: Filtering BLAST results using `filt_blast()` command**
+##### **Step 2: Filtering BLAST results using `filt_blast()` command**
 
[…]
