### 2020-08-19 #2473 @ramadatta
Hi @will-rowe - thank you. I am in the process of documenting of the package. Once I am done with this, I will start editing the final paper possibly this weekend and aim to submit soon.
