### 2020-08-22 622d6f7
Update README.md
added the example plots
