+++ README.md
@@ -170,7 +170,7 @@ svn export https://github.com/ramadatta/CPgeneProfiler/trunk/testData/fasta
 else simply [Click](https://downgit.github.io/#/home?url=https://github.com/ramadatta/CPgeneProfiler/tree/master/testData/fasta) to save fasta folder and uncompress the `fasta.zip` folder.
 
 
-##### **Step 1: A simple NCBI BLAST using `cpblast()` command**
+##### **1) A simple NCBI BLAST using `cpblast()` command**
 
 As a first step, CPgeneProfiler generates NCBI BLAST Results by aligning input genome assemblies against Carbapenamase (CP) gene database. Now that you already have a directory with fasta files (should have extensions `.fasta` or `.fa`) in `fasta` folder and cp gene database sequence in `db` folder, you can specify the path of both directories as an input and run the package with `cpblast()` command.
 
@@ -180,7 +180,7 @@ cpblast(fastalocation = "/home/user/CPgeneProfiler/testData/fasta", dblocation =
 The users can adjust BLAST parameters `num_threads`, `evalue`, `word_size` and `max_target_seqs` accordingly. If not adjusted and command is simply executed with the file locations for `fasta` and `db`, then default parameters are used for the analysis.
 
 
-##### **Step 2: Filtering BLAST results using `filt_blast()` command**
+##### **2) Filtering BLAST results using `filt_blast()` command**
 
 `filt_blast()` then filters the output BLAST results obtained from `cpblast()` command. This filtering is to find the presence of CP genes given a particular CP gene coverage and Percentage Identity. Therefore, the BLAST hits are filtered based on CP gene coverage and Percentage Identity. By default, CP Gene Coverage and Percentage Identity are set at a threshold of 100% (cpgcov=100, cpgpident=100). This means that a CP gene should have 100% alignment length and 100% identity, without even a single mismatch with the input genome sequence. The default parameters can be adjusted.
 
@@ -216,7 +216,7 @@ This should generate the following table:
 | genome_015.fasta | 3_length=66727_depth=0.76x                | OXA-181 | 66727  | 998  | 43441  | 44438  | 998  | 100 | 100 |
 | genome_015.fasta | 3_length=66727_depth=0.76x                | OXA-181 | 66727  | 998  | 37032  | 38029  | 998  | 100 | 100 |
 
-##### **Step 3c: Finding cocarriage genes using `cocarriage()` command**
+##### **3) Finding cocarriage genes using `cocarriage()` command**
 
 `cocarriage()` command finds if two or more CP genes exists in same contig or multiple contigs across all the input genome assemblies. This function can be used only after running `filt_blast()`. By default parameters, CP Gene Coverage and Percentage Identity are set to 100% (cpgcov=100, cpgpident=100) and can be adjusted.
 
@@ -224,7 +224,7 @@ This should generate the following table:
 cocarriage(cpgcov = 100, cpgpident = 100)
 ```
 
-##### **Step 3d: Finding CP gene profile using `cpprofile()` command**
+##### **4) Finding CP gene profile using `cpprofile()` command**
 
 `cpprofile()` creates a heatmap of CP gene profile across the input genome assemblies. By default, the command generates `png` image but user play with other output formats (jpeg/tiff/pdf) and parameters such as width, height of image, label, titles and colors of the heatmap.
 
@@ -235,7 +235,7 @@ cpprofile(outputType="png", width = 2000, height = 2000, res = 250, xlab="Carbap
 <img src="https://user-images.githubusercontent.com/3212461/90124487-10cca700-dd93-11ea-9572-dfc44a190dd6.png" width="45%"></img> 
 </p>
 
[…]
