+++ README.md
@@ -345,15 +345,19 @@ cp_summarize() | SummaryPlots.pdf | All the plots in a single pdf file
 ##### **Version**
 version 2.1.0
 
-#### Contributing and Contact Information
+# Community Guidelines
 
-We are open to suggestions for improvements in ``CPgeneProfiler``. Therefore, we invite all users to post a question or to provide us with (either positive or negative) feedback on the functions available in ``CPgeneProfiler``. When filing an issue, the most important thing is to include a minimal reproducible example so that we can quickly verify the problem, and then figure out how to fix it. There are two things you may need to include to make your example reproducible: data and code. If you are using additional packages, we would need that information as well. 
+## How to Contribute
 
-In addition, we welcome users who would like to contribute to the ``CPgeneProfiler`` package by adding new functions or co-developing some functions with us. You can let us know which function(s) you want to develop or which of the existing function(s) you want to improve. Please note that ``CPgeneProfiler`` is released with a [Contributor Code of Conduct](https://github.com/ramadatta/CPgeneProfiler/blob/master/Code_of_Conduct.md). By contributing to this project, you agree to abide by its terms.
+In general, you can contribute to this project by creating [issues](https://github.com/ramadatta/CPgeneProfiler/issues).
+You are also welcome to contribute to the source code directly by forking the project, modifying the code, and creating [pull requests](https://github.com/ramadatta/CPgeneProfiler/pulls).
+If you are not familiar with pull requests, check out [this post](https://guides.github.com/activities/forking/).
+Please use clear and organized descriptions when creating issues and pull requests.
 
+Please note that ``CPgeneProfiler`` is released with a [Contributor Code of Conduct](https://github.com/ramadatta/CPgeneProfiler/blob/master/Code_of_Conduct.md). By contributing to this project, you agree to abide by its terms.
 
-To get further help regarding the functions available in ``CPgeneProfiler`` or inquries regarding contributions, please email:
+## Bug Report and Support Request
 
-  - [Prakki Sai Rama Sridatta](https://twitter.com/prakki_rama) (<ramadatta.88@gmail.com>)
-  
+You can use [issues](https://github.com/ramadatta/CPgeneProfiler/issues) to report bugs and seek support.
+Before creating any new issues, please check for similar ones in the issue list first. 
 
