+++ README.md
@@ -42,7 +42,7 @@ Generate a profile of carbapenamase genes from the genome assemblies
 	 pdftools,
 	 grid
 	 
-Install these packages using R (>=3.5):
+Install these packages using R (versions 3.5 & 3.6):
 	
 ``` r
 install.packages(c("BiocManager", "tidyverse", "UpSetR", "scales", "ape", 
