+++ R/db_summary.R
@@ -12,11 +12,11 @@
 db_summary <- function()
 {
   print("DATABASE: NCBI Bacterial Antimicrobial Resistance Reference Gene Database")
-  print("VERSION: 2020-06-11.1")
-  print("SEQUENCES: 1062")
+  print("VERSION: 2020-07-16.2")
+  print("SEQUENCES: 875")
   print("DBTYPE: nucl")
-  print("DATE: 2020-Aug-04")
-  print("Reference Gene Catalog: https://www.ncbi.nlm.nih.gov/pathogens/isolates#/refgene/")
+  print("DATE: 2020-Aug-23")
+  print("Reference Gene Catalog: ftp://ftp.ncbi.nlm.nih.gov/pathogen/Antimicrobial_resistance/AMRFinderPlus/database/3.8/")
 }
 
 
