### 2020-09-03 #2473 @ramadatta
Hi @will-rowe - just an update on manuscript. I am almost done with manuscript and awaiting my collaborator's go-ahead to submit. Thank you.
