### 2020-09-28 ramadatta/CPgeneProfiler#14 @ramadatta
We completely agree with reviewer’s comment. Based on the reviewer's recommendation, we have replaced ARG-ANNOT CP gene database with a CP gene database extracted from a very recent version of NCBI Bacterial Antimicrobial Resistance Reference Gene Database (version 2020-07-16.2). The criteria on how the CP genes were selected from the NCBI database and the exclusion of non-carbapenemase OXA gene variants are also now mentioned in the manuscript.
