### 2020-09-28 ramadatta/CPgeneProfiler#12 @ramadatta
We completely agree with reviewer’s comment on splitting single function to multiple to improve utility. `CPgeneProfiler` now has separate documented functions such as  `cpblast`, `filt_blast` which deals NCBI BLAST tool results and other visualization commands such as `cpprofile`, `plot_conlen`, `upsetR`. For all these functions, users can now tweak the parameters such as percent identity, alignment coverage, various plot sizes and colors. The [example usage](https://github.com/ramadatta/CPgeneProfiler#example-usage) section details each of these functions.

But given the time constraints, we could not specifically create a upset plot function for plotting specific CP gene/families. We surely consider to make this available in the upcoming versions of the package.
