### 2020-09-28 ramadatta/CPgeneProfiler#9 @ramadatta
This issue has been fixed now. The CP gene profile [output plot](https://github.com/ramadatta/CPgeneProfiler/blob/master/testData_Output/3_CP_Gene_Profile/CPgeneProfile.png) now has genome assembly names. The color label is also now reset to “Gene Count”.
