### 2020-09-28 ramadatta/CPgeneProfiler#10 @ramadatta
This issue has been fixed. The plot now appears properly without any blank pages. Please find the link: https://github.com/ramadatta/CPgeneProfiler/blob/master/testData_Output/6_UpsetR_Intersection_Plot/upset_plot.pdf
