### 2020-09-28 ramadatta/CPgeneProfiler#7 @ramadatta
We agree with reviewer's comment. These issues have been fixed.

R package dependencies are added to [DESCRIPTION](https://github.com/ramadatta/CPgeneProfiler/blob/master/DESCRIPTION) file. We are hopeful that this would ensure smooth installation of the package using devtools.

Additionally, in the README page under the [Requirements](https://github.com/ramadatta/CPgeneProfiler#requirements) section, we have provided the R commands to install relevant dependencies.

 At this juncture, in view of submission time lines, we could not readily come up with bioconda package. But, we will surely consider making a bioconda package with all the dependencies in the upcoming versions of the package.
