### 2020-09-28 ramadatta/CPgeneProfiler#5 @ramadatta
This issue has been fixed. DESCRIPTION metadata file now has authorship and relevant information. Please find the link here: https://github.com/ramadatta/CPgeneProfiler/blob/master/DESCRIPTION
