### 2020-09-28 ramadatta/CPgeneProfiler#19 @ramadatta
Our initial intention was to include the appropriate journal references in the README page. But, after exploring README page format of some of the JOSS accepted papers more closely, it appears that there is no specific requirement from the journal to have references in the README page. We admit that our initial list of references was partial which happened due to our oversight. We now have taken away the **References** section from documentation and listed all the relevant references in the manuscript.
