### 2020-09-28 ramadatta/CPgeneProfiler#11 @ramadatta
We agree with the reviewer's comment on having testing section. In accordance to that, we now have “[Working with testData](https://github.com/ramadatta/CPgeneProfiler#working-with-testdata)” section which give details of each command and displays the expected outcome.

But given the time constraints, we could not automate the testing process. We will surely consider to do this upcoming versions of the package.
