### 2020-09-28 ramadatta/CPgeneProfiler#13 @ramadatta
This issue has been fixed. We have now added community guidelines in our README page. This would help users understand on how to contribute, support and report any issues with software. Please find the link here: https://github.com/ramadatta/CPgeneProfiler#community-guidelines
