### 2020-09-28 ramadatta/CPgeneProfiler#23 @ramadatta
We agree with reviewer’s comment on this. The rewritten package now has commands such as `filt_blast`, `cocarriage`, `plot_conlen`, `upsetR_plot`. All these commands now has an option to filter based on `cpgpident` (nucleotide identity) and `cpgcov` (coverage) in the interest of the user. The default settings have been documented for each of these commands. Please find the [example usage](https://github.com/ramadatta/CPgeneProfiler#example-usage).
