### 2020-09-29 ramadatta/CPgeneProfiler#22 @ramadatta
We agree with reviewer’s comment and this issue has been resolved. The output directory could now be specified using `cp_summarize` command.

For example:
`cp_summarize(outdir_loc = "/home/user/test", outdir = "xyz_directory", report="SummaryPlots" , image = "png")`

The above command will automatically create a `xyz_directory` folder under `/home/user/test` location and transfers all the output files into `xyz_directory` at the specified location. Please find the details of `cp_summarize` command from the [testData](https://github.com/ramadatta/CPgeneProfiler#working-with-testdata) section.
