### 2020-09-29 ramadatta/CPgeneProfiler#21 @ramadatta
We agree with reviewer’s comment on generating a summary report. Accordingly, all the output plots generated from various commands are summarized now in a single PDF file. This can be achieved by running `cp_summarize` command. An example output of the summary plots can be found [here](https://github.com/ramadatta/CPgeneProfiler/blob/master/testData_Output/SummaryPlots.pdf).

Additionally, `cp_summarize` command arranges all the output files from different commands into respective folders based on the analysis performed. This should resolve the issue with scattered output data files.

For example, blast results and filtered blast results are placed under the folder name "[1_CP_BLAST_Results](https://github.com/ramadatta/CPgeneProfiler/tree/master/testData_Output/1_CP_BLAST_Results)" and all the cocarriage output files are placed under the folder name "[2_Cocarriage_Results](https://github.com/ramadatta/CPgeneProfiler/tree/master/testData_Output/2_Cocarriage_Results)".

Also, as per the reviewer’s comment, the [cocarriage analysis results](https://github.com/ramadatta/CPgeneProfiler/blob/master/testData_Output/2_Cocarriage_Results/Cocarriage_combinedResults.txt) will now contain all the information such as contig name, contig position of the CP gene, contig length, co-occurring CP genes, blast results, coverage, nucleotide identity, e-value.
