### 2020-09-29 ramadatta/CPgeneProfiler#24 @ramadatta
We completely agree with reviewer’s comment. Accordingly based on other reviewer's comment, we have now replaced ARG-ANNOT CP gene database with a CP gene database extracted from a very recent version of NCBI Bacterial Antimicrobial Resistance Reference Gene Database (2020-07-16.2). This database is continuously updated and delineates the CP genes by categorizing the genes with carbapenemase activity as "CARBAPENEM"  in the [subclass](https://www.ncbi.nlm.nih.gov/pathogens/pathogens_help/#refgene-data-field-subclass). Therefore, only those genes whose subclass was labelled as “CARBAPENEM” in the reference gene catalog were considered for database preparation. This criteria ensures to ignore genes which are devoid of carbapenemase activity (for example, OXA variants such as OXA-163 and OXA-405 genes).
