### 2020-09-29 ramadatta/CPgeneProfiler#25 @ramadatta
We agree with the reviewer. The default parameters are mentioned in the documentation now. Users can now tweak the alignment coverage and percentage identity parameters while running `filt_blast` and `cocarriage` commands. Please find the details in the [example usage](https://github.com/ramadatta/CPgeneProfiler#example-usage) section.
