### 2020-10-05 6b0e891
Delete paper.md
