### 2020-10-05 7e1aa51
Delete paper.bib
