+++ paper.md
@@ -1,0 +1,131 @@
+---
+title: '  CPgeneProfiler: A lightweight R package to profile the Carbapenamase genes
+  from genome assemblies'
+tags:
+- R
+- Carbapenamase gene profile
+- cocarriage
+- Beta-lactamases
+- Antimicrobial resistance (AMR)
+date: "13 June 2020"
+output:
+  rticles::joss_article: default
+  pdf_document: default
+  word_document: default
+authors:
+- name: Prakki Sai Rama Sridatta
+  orcid: 0000-0002-9254-2557
+  affiliation: '1,2'
+- name: Natascha M Thevasagayam
+  affiliation: '1,2'
+- name: Weizhen Xu
+  affiliation: '1,2'
+- name: Kalisvar Marimuthu
+  affiliation: '1,2'
+- name: Jeanette W P Teo
+  affiliation: 3
+- name: Indumathi Venkatachalam
+  affiliation: 4
+- name: Ng Oon Tek
+  affiliation: '1,2,5'
+year: 2020
+bibliography: paper.bib
+citation_author: Prakki et. al.
+csl: apa.csl
+journal: JOSS
+affiliations:
+- name: National Centre for Infectious Diseases, Singapore
+  index: 1
+- name: Tan Tock Seng Hospital, Singapore
+  index: 2
+- name: National University Hospital, Singapore
+  index: 3
+- name: Singapore General Hospital, Singapore
+  index: 4
+- name: Lee Kong Chian School of Medicine, Nanyang Technological University, Singapore
+  index: 5
+---
+
+# Summary
+
+“Carbapenems” are a specific subset of antibiotics considered to possess a higher spectrum of antimicrobial activity [@papp2011carbapenems] against Gram-positive and Gram-negative bacteria. Even so, there are pathogens which are resistant to carbapenems due to the presence of carbapenemase genes (CP genes) which have the ability to hydrolyze carbapenems. 
+
+Studies show that those cases infected by carbapenem-resistant pathogens have a higher morbidity and mortality rate compared with those who are infected by non-carbapenem-resistant pathogens [@van2013carbapenem; @cai2017prevalence]. Therefore, early discerning of the CP genes and their resistance mechanisms are considered crucial to aid in infection control as well as lessen the likelihood of mortality, duration of hospitalization stay, and related medical costs [@van2013carbapenem; @nordmann2019epidemiology]. Further, it is understood that the cocarriage of genes encoding different classes of carbapenemases could confer higher resistance to carbapenem antibiotics, which may promote further spread of the disease [@wang2019cocarriage].
+
+The detection of the resistance genes from various bacterial strains using techniques such as polymerase chain reaction (PCR) and microarrays in real-time are very time consuming and costly. With the advancement in whole-genome sequencing (WGS) technologies the costs are more accessible and WGS provides an alternative method for detection of resistance genes, given that the relevant analysis tools are available.
+
[…]
