+++ paper.bib
@@ -1,0 +1,289 @@
+@article{panunzi2020srax,
+  title={sraX: A Novel Comprehensive Resistome Analysis Tool},
+  author={Panunzi, Leonardo G},
+  journal={Frontiers in microbiology},
+  volume={11},
+  pages={52},
+  year={2020},
+  publisher={Frontiers},
+  doi={10.3389/fmicb.2020.00052}
+}
+
+@article{alcock2020card,
+  title={CARD 2020: antibiotic resistome surveillance with the comprehensive antibiotic resistance database},
+  author={Alcock, Brian P and Raphenya, Amogelang R and Lau, Tammy TY and Tsang, Kara K and Bouchard, M{\'e}gane and Edalatmand, Arman and Huynh, William and Nguyen, Anna-Lisa V and Cheng, Annie A and Liu, Sihan and others},
+  journal={Nucleic acids research},
+  volume={48},
+  number={D1},
+  pages={D517--D525},
+  year={2020},
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkz935}
+}
+
+@article{zankari2017pointfinder,
+  title={PointFinder: a novel web tool for WGS-based detection of antimicrobial resistance associated with chromosomal point mutations in bacterial pathogens},
+  author={Zankari, Ea and Alles{\o}e, Rosa and Joensen, Katrine G and Cavaco, Lina M and Lund, Ole and Aarestrup, Frank M},
+  journal={Journal of Antimicrobial Chemotherapy},
+  volume={72},
+  number={10},
+  pages={2764--2768},
+  year={2017},
+  publisher={Oxford University Press},
+  doi={10.1093/jac/dkx217}	
+}
+
+@article{bortolaia2020resfinder,
+  title={ResFinder 4.0 for predictions of phenotypes from genotypes},
+  author={Bortolaia, Valeria and Kaas, Rolf S and Ruppe, Etienne and Roberts, Marilyn C and Schwarz, Stefan and Cattoir, Vincent and Philippon, Alain and Allesoe, Rosa L and Rebelo, Ana Rita and Florensa, Alfred Ferrer and others},
+  journal={Journal of Antimicrobial Chemotherapy},
+  year={2020},
+  doi={10.1093/jac/dkaa345}
+}
+
+
+@article{clausen2016benchmarking,
+  title={Benchmarking of methods for identification of antimicrobial resistance genes in bacterial whole genome data},
+  author={Clausen, Philip TLC and Zankari, Ea and Aarestrup, Frank M and Lund, Ole},
+  journal={Journal of Antimicrobial Chemotherapy},
+  volume={71},
+  number={9},
+  pages={2484--2488},
+  year={2016},
+  publisher={Oxford University Press},
+  doi={10.1093/jac/dkw184}	
+}
+
+@article{clausen2018rapid,
+  title={Rapid and precise alignment of raw reads against redundant databases with KMA},
+  author={Clausen, Philip TLC and Aarestrup, Frank M and Lund, Ole},
+  journal={BMC bioinformatics},
+  volume={19},
+  number={1},
+  pages={1--8},
+  year={2018},
+  publisher={BioMed Central},
+  doi={10.1186/s12859-018-2336-6}	
+}
+
+@article{conway2017upsetr,
+  title={UpSetR: an R package for the visualization of intersecting sets and their properties},
+  author={Conway, Jake R and Lex, Alexander and Gehlenborg, Nils},
+  journal={Bioinformatics},
+  volume={33},
+  number={18},
+  pages={2938--2940},
+  year={2017},
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btx364}
+}
+
+@Book{ggplot,
[…]
