### 2020-10-05 ramadatta/CPgeneProfiler#6 @ramadatta
We agree with reviewer on the conflation of AMR databases and AMR gene detection tools in the manuscript. In the updated manuscript, we have rephrased the paragraph by mentioning only about the relevant AMR gene detection tools but not the databases.
