### 2020-10-05 ramadatta/CPgeneProfiler#16 @ramadatta
We agree with reviewer on comparison to AMR detection tools with summary utilities. In regard to that, we have rephrased in the manuscript how other AMR detection tools are undeniably useful in generating summary tables and visualizations with full AMR complement. And also, we have mentioned on the benefits of the specialization of CPgeneProfiler tool.
