### 2020-10-05 a647150
Delete paper.md
