### 2020-10-05 ramadatta/CPgeneProfiler#17 @ramadatta
We agree with reviewer on this. In the updated manuscript we have elaborated the usefulness of each of the utility in relation to Carbapenemase (CP) gene analysis.
