### 2020-10-05 ramadatta/CPgeneProfiler#18 @ramadatta
We agree with reviewer on this. In the updated manuscript, we have a paragraph on how the CP gene database is created. We have clarified in that section that some of the OXA variants (such as OXA-163 and OXA-405) are not carbapenemases and such gene variants were excluded from the CP gene database.
