+++ paper.md
@@ -54,9 +54,7 @@ Studies show that those cases infected by carbapenem-resistant pathogens have a 
 
 The detection of the resistance genes from various bacterial strains using techniques such as polymerase chain reaction (PCR) and microarrays in real-time are very time consuming and costly. With the advancement in whole-genome sequencing (WGS) technologies the costs are more accessible and WGS provides an alternative method for detection of resistance genes, given that the relevant analysis tools are available.
 
-To this end, several freely available bioinformatics tools such as ``ABRicate`` (https://github.com/tseemann/abricate), ``AMRPlusPlus`` [@doster2020megares], ``ARG-ANNOT`` [@gupta2014arg], ``ARIBA`` [@hunt2017ariba], ``Comprehensive Antibiotic Resistance Database – Resistance Gene Identifier`` (CARD-RGI) [@alcock2020card], ``NCBI AMRFinderPlus`` (https://ncbi.nlm.nih.gov/pathogens/antimicrobial-resistance/AMRFinder/), ``KmerResistance`` [@clausen2016benchmarking; @clausen2018rapid], ``PointFinder`` [@zankari2017pointfinder], ``Resfinder`` [@bortolaia2020resfinder], ``sraX`` [@panunzi2020srax] and ``SRST2`` [@inouye2014srst2] assist in finding of the antimicrobial resistance genes from the sequence data [@hendriksen2019using]. Among these, tools such as ``ABRicate`` and ``CARD-RGI`` can generate comparative tables across genomes and ``sraX`` can help in visualization of comprehensive AMR gene complement. 
-
-However, in the research settings where it is crucial to quickly examine the transmission of CP genes, it is useful to have a tool that is catered to CP gene dataset that provides easily interpretable visualizations and statistics. Undeniably, all the above-mentioned tools are focused around all the antimicrobial-resistant genes and tools such as ABRicate and CARD-RGI can even generate comparative tables across genomes and sraX can help in visualization of comprehensive AMR gene complement. But to our knowledge, all these tools do not solely provide a genetic profile for the presence of CP genes across multiple genome assemblies, and neither do they extract and visualize the cocarriage of multiple CP genes across the genome assemblies. Therefore, to address this need, we describe here a lightweight R package, **CPgeneProfiler** that scans multiple bacterial genome assemblies to detect and visualize the presence of CP genes using the R framework.
[…]
