+++ paper.md
@@ -56,7 +56,7 @@ The detection of the resistance genes from various bacterial strains using techn
 
 To this end, several freely available bioinformatics tools such as ``ABRicate`` (https://github.com/tseemann/abricate), ``AMRPlusPlus`` [@doster2020megares], ``ARG-ANNOT`` [@gupta2014arg], ``ARIBA`` [@hunt2017ariba], ``Comprehensive Antibiotic Resistance Database – Resistance Gene Identifier`` (CARD-RGI) [@alcock2020card], ``NCBI AMRFinderPlus`` (https://ncbi.nlm.nih.gov/pathogens/antimicrobial-resistance/AMRFinder/), ``KmerResistance`` [@clausen2016benchmarking; @clausen2018rapid], ``PointFinder`` [@zankari2017pointfinder], ``Resfinder`` [@bortolaia2020resfinder], ``sraX`` [@panunzi2020srax] and ``SRST2`` [@inouye2014srst2] assist in finding of the antimicrobial resistance genes from the sequence data [@hendriksen2019using]. However, in the research settings where it is crucial to quickly examine the transmission of CP genes, it is useful to have a tool that is catered to CP gene dataset that provides easily interpretable visualizations and statistics. Undeniably, all the above-mentioned tools are focused around the antimicrobial-resistant genes and tools such as ABRicate and CARD-RGI can even generate comparative tables across genomes and sraX can help in visualization of comprehensive AMR gene complement. But to our knowledge, all these tools do not solely provide a genetic profile for the presence of CP genes across multiple genome assemblies, and neither do they extract and visualize the cocarriage of multiple CP genes across the genome assemblies. Therefore, to address this need, we describe here a lightweight R package, **CPgeneProfiler** that scans multiple bacterial genome assemblies to detect and visualize the presence of CP genes using the R framework.
 
-In order to detect CP genes from the genome assemblies, ``NCBI Bacterial Antimicrobial Resistance Reference Gene Database`` (2020-07-16.2) (https://www.ncbi.nlm.nih.gov/bioproject/PRJNA313047) was used for generating a CP gene database. Only those genes whose subclass is categorized as "CARBAPENEM" in the reference gene catalog were considered for database preparation. This excluded the possibility of having resistant gene variants which are beta-lactamases but do not show carbapenem-resistant activity. For example, although OXA-48 is a carbapenemase gene, it's variants such as OXA-163 and OXA-405 have been concluded to be devoid of any carbapenemase activity [@dortet2017noncarbapenemase] and therefore were not considered being part of the CP gene database. 
+In order to detect CP genes from the genome assemblies, ``NCBI Bacterial Antimicrobial Resistance Reference Gene Database`` (2020-07-16.2) (https://www.ncbi.nlm.nih.gov/bioproject/PRJNA313047) was used for generating a CP gene database. Only those resistant genes whose subclass is categorized as "CARBAPENEM" in the reference gene catalog were considered for database preparation. This excluded the possibility of having resistant gene variants which are beta-lactamases but do not show carbapenem-resistant activity. For example, although OXA-48 is a carbapenemase gene, other OXA variants such as OXA-163 and OXA-405 have been concluded to be devoid of any carbapenemase activity [@dortet2017noncarbapenemase] and therefore their subclass is not categorized as "CARBAPENEM" in the NCBI Bacterial Antimicrobial Resistance Reference Gene Database. Therefore, both OXA variants OXA-163, OXA-405 were not included in the CP gene database.
 
[…]
