### 2020-10-06 ramadatta/CPgeneProfiler#15 @ramadatta
We agree with reviewer's comment on this. Accordingly, we have now cited the latest manuscript references for all the tools mentioned in the manuscript.
