### 2020-10-08 #2473 @will-rowe
Hi @ramadatta. Many thanks for your thorough responses to the comments. In my opinion you have satisfied the reviewer comments. That being said, if @fmaguire or @boasvdp would like to give their thoughts on your responses, that would be most welcome.

I have a request and a question:

* Please could you add a specific **statement of need** section to the paper. I think you describe the need within the paper but it would be better if this was explicit. [here](https://joss.theoj.org/papers/10.21105/joss.02451) is an example paper layout which includes the statement of need section.

* I noticed a PR go in to BioConda a few weeks ago for a CPgeneProfiler recipe but I can't see it in the package list. Was there any roadblock to getting this in? You don't need it in BioConda for us to accept the paper but I think it would be really good to have it. I can try and help you get it packaged if there is a problem, or we can ask on the Bioconda team gitter for assistance.
