### 2020-10-09 #2473 @fmaguire
I've updated the checklist in light of the changes. Only minor issue is that when @ramadatta makes the explicit "statement of need" section in the paper it could do with being added to the README as well.

Apart from that, all good from my end, @ramadatta did a great job thoroughly addressing the comments. The tool and its documentation is much improved from their work.
