### 2020-10-09 #2473 @ramadatta
Dear @will-rowe - thank you very much for your comment.

Thanks for providing the example paper for adding statement of need section in paper. We will edit our manuscript accordingly.

We do agree, it is very helpful for users to install of package through conda. In regards to this, we did try to package the [recipe](https://github.com/openjournals/joss-reviews/files/5352239/meta_yaml.txt). But for some reason, I had trouble getting this in [Bioconda](https://github.com/bioconda/bioconda-recipes/pull/24047).

We sincerely appreciate and are thankful to you for offering help to get it packaged. At the moment, I have [meta.yaml](https://github.com/openjournals/joss-reviews/files/5352239/meta_yaml.txt) file and if possible could I kindly request to get pointers from you to work on packaging this.
