### 2020-10-09 #2473 @fmaguire
From the bioconda linter
(https://app.circleci.com/pipelines/github/bioconda/bioconda-recipes/31534/workflows/2aeaba6e-e261-4314-b570-17b5cb046873/jobs/120863),
the following seem to be the issues:

```ERROR: recipes/cpgeneprofiler/meta.yaml:11:
should_be_noarch_generic: The recipe should be build as `noarch`
ERROR: recipes/cpgeneprofiler/meta.yaml:2:
folder_and_package_name_must_match: The recipe folder and package name
do not match.
WARNING: recipes/cpgeneprofiler/meta.yaml:28:
gpl_requires_license_distributed: The recipe packages GPL software but
is missing copy of license.
ERROR: recipes/cpgeneprofiler/meta.yaml:9: build_number_needs_reset:
The recipe build number should be reset to 0```
