### 2020-10-09 #2473 @boasvdp
Hi @ramadatta, all issues have been addressed appropriately in my opinion. I have updated my checklist.

I agree adding this to bioconda would be great!
