### 2020-10-09 #2473 @ramadatta
Dear @will-rowe @fmaguire - Thank you. We have now made the statement of purpose more explicit in the paper and documentation. Please kindly review our updated manuscript and [synopsis](https://github.com/ramadatta/CPgeneProfiler/blob/master/README.md#synopsis) section in documentation.

@fmaguire @boasvdp - I will restart working on conda packaging. Thank you for the advice.
