### 2020-10-09 #2473 @will-rowe
Thanks @ramadatta - this looks good. I think we can move toward acceptance!

If you can now tag a new release, archive it (zenodo) and then report back here with the version and DOI.

I've got the recipe building (thanks @fmaguire for the linting report). I will wait for your new release before we make the bioconda PR
