### 2020-10-10 #2473 @ramadatta
Dear @will-rowe - I have tag for the [CPgeneProfiler v2.1.1](https://github.com/ramadatta/CPgeneProfiler/tree/v2.1.1) and the DOI is: [10.5281/zenodo.4076089](https://zenodo.org/record/4076089)
