+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: CPgeneProfiler
 Title: Profiles the Carbapenamase genes from the genome assemblies (One Line, Title Case)
-Version: 2.1.1
+Version: 2.1.0.9000
 Authors@R: 
     c(person(given = "Sai Rama Sridatta",
            family = "Prakki",
