### 2020-10-11 #2473 @will-rowe
Sorry - missed the question re. editing. Sure you can - I will wait before flagging this paper to the editor in chief. Let me know when you are ready for the paper to be accepted. You will need another release/DOI if you make edits, so let me know what they are.
