### 2020-10-11 #2473 @will-rowe
Bioconda recipe is ready to be reviewed and merged (see [here](https://github.com/bioconda/bioconda-recipes/pull/24782)). I had to add the dependencies to the meta.yaml, add the license and sort a name conflict. I also set the recipe to build from your releases, rather than the master branch, which I think is preferred by Bioconda.
