### 2020-10-12 #2473 @ramadatta
> Sorry - missed the question re. editing. Sure you can - I will wait before flagging this paper to the editor in chief. Let me know when you are ready for the paper to be accepted. You will need another release/DOI if you make edits, so let me know what they are.

Thank you very much @will-rowe - I just made a minor [edit](https://github.com/ramadatta/CPgeneProfiler/commit/0f20d3b746f07e86d70b25d3575bc9052d5e9409#diff-fb6416b20e925ad59e9f77c28b957d88) at **`visualize the set intersections of cocarriage of CP genes`** in the manuscript.

Also, as you suggested, I have updated the CPgeneProfiler [v2.1.1](https://github.com/ramadatta/CPgeneProfiler/compare/v2.1.1...master) release and updated the Zenodo repo [author list](https://zenodo.org/record/4076089). Thank you so much for the Bioconda request.

Please let me know if I missed anything else from side. If nothing, I kindly request the editor for the manuscript to accepted.
