### 2020-10-12 dd13bc5
wording changes in paper
