### 2020-10-12 #2473 @danielskatz
Hi - I'm the JOSS AEiC this week, and I'll be finishing the processing of your submission.  There are some wording problems in the paper - please either merge https://github.com/ramadatta/CPgeneProfiler/pull/26 or let me know what you disagree with.  In addition, a number of the references have case issues - please be sure all journals are in Title Case - use {}s in the bib file to protect cases as needed.
