### 2020-10-12 #2473 @ramadatta
Dear @danielskatz - Thank you very much!

All the edits you made are completely acceptable for us, except the following:

In line 69: Following the terminology of the [paper](https://academic.oup.com/bioinformatics/article/33/18/2938/3884387), we want it to be: **set intersections**

In line 73: N50, N90 and assembly size are 3 different statistics. So, it should be:
**``CPgeneProfiler`` can also generate the N50, N90 and assembly size statistics for each of the genome assemblies.**

We did a careful recheck about the case issues of the tools that were referred in the manuscript. We are confident that we are following the citation as the tool authors required it to be cited.

Please let us know if we are missing anything. Thanks much!
