+++ paper.md
@@ -66,7 +66,7 @@ In order to detect CP genes from the genome assemblies, ``NCBI Bacterial Antimic
 
 The tool first uses the ``cpblast`` command, by which each fasta file is searched against the CP gene database using NCBI BLAST+ [@camacho2009blast] (version 2.9.0+), which is pre-installed in the local system as a dependency. The presence of a CP gene in an assembled genome is confirmed if the CP gene meets the identity and coverage thresholds (default: 100%) when aligned with the genome sequence. The genome sequences that meet the thresholds are extracted from the BLAST results using the ``filt_blast`` command.
 
-Visualizing the presence of CP genes and their corresponding counts across all the genome assemblies in a simple heatmap enables one to find CP gene variants that are found across the samples and aids in exploring the pattern of CP gene presence with reference to species or sequence type (ST). In order to facilitate this, the ``cpprofile`` command generates a profile of CP genes (Figure 1A) across the genome assemblies, while the ``cocarriage`` command finds cocarriage of CP genes in the genome assemblies. In addition to this, the tool also generates plots to visualize the set of intersections of CP genes across all the input genome assemblies using the command ``upsetR_plot`` (Figure 1B). It is understood that isolates that harbour multiple carbapenemase genes are considered to produce high resistant phenotypes, and running the commands ``cocarriage`` and ``upsetR_plot`` provides an overview of the CP genes as well as their cocarriages present in all the genomes.
+Visualizing the presence of CP genes and their corresponding counts across all the genome assemblies in a simple heatmap enables one to find CP gene variants that are found across the samples and aids in exploring the pattern of CP gene presence with reference to species or sequence type (ST). In order to facilitate this, the ``cpprofile`` command generates a profile of CP genes (Figure 1A) across the genome assemblies, while the ``cocarriage`` command finds cocarriage of CP genes in the genome assemblies. In addition to this, the tool also generates plots to visualize the set intersections of CP genes across all the input genome assemblies using the command ``upsetR_plot`` (Figure 1B). It is understood that isolates that harbour multiple carbapenemase genes are considered to produce high resistant phenotypes, and running the commands ``cocarriage`` and ``upsetR_plot`` provides an overview of the CP genes as well as their cocarriages present in all the genomes.
 
 Given a set of bacterial genomes that are of same species, it would be useful to explore if the CP genes are found on specific plasmids or scattered across multiple plasmids/chromosomes of different sequence lengths. This can be achieved by plotting the number of contigs across the contig length by using the ``plot_conlen`` command (Figure 2). 
 
