### 2020-10-12 #2473 @danielskatz
I've changed the 2 things you disagreed with - please see if this is now ok, and if so, please merge.
