### 2020-10-12 3230467
Update paper.bib
changes to cases in journal titles
