+++ paper.bib
@@ -1,7 +1,7 @@
 @article{panunzi2020srax,
   title={sraX: A Novel Comprehensive Resistome Analysis Tool},
   author={Panunzi, Leonardo G},
-  journal={Frontiers in microbiology},
+  journal={Frontiers in Microbiology},
   volume={11},
   pages={52},
   year={2020},
@@ -12,7 +12,7 @@
 @article{alcock2020card,
   title={CARD 2020: antibiotic resistome surveillance with the comprehensive antibiotic resistance database},
   author={Alcock, Brian P and Raphenya, Amogelang R and Lau, Tammy TY and Tsang, Kara K and Bouchard, M{\'e}gane and Edalatmand, Arman and Huynh, William and Nguyen, Anna-Lisa V and Cheng, Annie A and Liu, Sihan and others},
-  journal={Nucleic acids research},
+  journal={Nucleic Acids Research},
   volume={48},
   number={D1},
   pages={D517--D525},
@@ -57,7 +57,7 @@
 @article{clausen2018rapid,
   title={Rapid and precise alignment of raw reads against redundant databases with KMA},
   author={Clausen, Philip TLC and Aarestrup, Frank M and Lund, Ole},
-  journal={BMC bioinformatics},
+  journal={BMC Bioinformatics},
   volume={19},
   number={1},
   pages={1--8},
@@ -90,7 +90,7 @@
 @article{camacho2009blast,
   title={BLAST+: architecture and applications},
   author={Camacho, Christiam and Coulouris, George and Avagyan, Vahram and Ma, Ning and Papadopoulos, Jason and Bealer, Kevin and Madden, Thomas L},
-  journal={BMC bioinformatics},
+  journal={BMC Bioinformatics},
   volume={10},
   number={1},
   pages={421},
@@ -102,7 +102,7 @@
 @article{inouye2014srst2,
   title={SRST2: rapid genomic surveillance for public health and hospital microbiology labs},
   author={Inouye, Michael and Dashnow, Harriet and Raven, Lesley-Ann and Schultz, Mark B and Pope, Bernard J and Tomita, Takehiro and Zobel, Justin and Holt, Kathryn E},
-  journal={Genome medicine},
+  journal={Genome Medicine},
   volume={6},
   number={11},
   pages={90},
@@ -114,7 +114,7 @@
 @article{zankari2012identification,
   title={Identification of acquired antimicrobial resistance genes},
   author={Zankari, Ea and Hasman, Henrik and Cosentino, Salvatore and Vestergaard, Martin and Rasmussen, Simon and Lund, Ole and Aarestrup, Frank M and Larsen, Mette Voldby},
-  journal={Journal of antimicrobial chemotherapy},
+  journal={Journal of Antimicrobial Chemotherapy},
   volume={67},
   number={11},
   pages={2640--2644},
@@ -126,7 +126,7 @@
 @article{lakin2017megares,
   title={MEGARes: an antimicrobial resistance database for high throughput sequencing},
   author={Lakin, Steven M and Dean, Chris and Noyes, Noelle R and Dettenwanger, Adam and Ross, Anne Spencer and Doster, Enrique and Rovira, Pablo and Abdo, Zaid and Jones, Kenneth L and Ruiz, Jaime and others},
-  journal={Nucleic acids research},
+  journal={Nucleic Acids Research},
   volume={45},
   number={D1},
   pages={D574--D580},
@@ -138,7 +138,7 @@
 @article{doster2020megares,
   title={MEGARes 2.0: a database for classification of antimicrobial drug, biocide and metal resistance determinants in metagenomic sequence data},
[…]
