### 2020-10-12 #2473 @danielskatz
Re cases, I've just gone ahead with https://github.com/ramadatta/CPgeneProfiler/pull/27 - please merge this as well, or again, let me know what you disagree with
