### 2020-10-12 #2473 @ramadatta
> Re cases, I've just gone ahead with [ramadatta/CPgeneProfiler#27](https://github.com/ramadatta/CPgeneProfiler/pull/27) - please merge this as well, or again, let me know what you disagree with

Thank you @danielskatz - I have merged both the updates. Please let us know if anything else.
