### 2020-10-13 #2473 @ramadatta
We ate thoroughly impressed with JOSS editors and review process.

We are very thankful to @fmaguire @boasvdp and @will-rowe for comprehensive reviews and guidance.
